[Playground] Missing playground results for property accessors in a class wrapped in a struct
When a decl that has properties with accessors (`get`, `set`, `willSet`, `didSet`, etc) is nested inside another type, those accessor implementations aren't playground-transformed.

The reason is that the playground transform uses an `ASTWalker` to get to the top-level structure, but then once it’s inside a type, it does directly nested `transformDecl()` calls. And that inner check is missing a case for accessors.

This change adds an else-clause that checks if the declaration is a `VarDecl`, and if so, calls `transformDecl()` on each accessor.

It's unfortunate that the playground transform reaches decls in two different ways (using an `ASTWalker` to get to the top-level decls but then using directly nested calls below). This seems to be worth resolving at some point (perhaps by using the `ASTWalker` for the whole traversal?), but that’s a larger change, and so this is worth addressing with a safer short-term fix.

rdar://139656464